+++ ts/src/index.ts
@@ -1,5 +1,6 @@
 export * from './dna_factory';
 export * from './eggs_factory';
+export * from './dna_schema_reader';
 export * from './interfaces/types';
 export * from './constants';
 export * as utils from './utils';